# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ProyectoTech",
        "description": "Documentación proyectoTech (CRUD de productos y carga masiva - Carrito de Compras)",
        "contact": {
            "email": "johan33@hotmail.es"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/",
            "description": "Desarrollo ProyectoTech"
        }
    ],
    "paths": {
        "/auth/usuarios/registro": {
            "post": {
                "tags": [
                    "Usuarios - Autenticación"
                ],
                "summary": "Nuevo usuario",
                "description": "Servicio para el registro de un nuevo usuario",
                "operationId": "usuarios/registro",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nombre del usuario a registrar",
                                        "type": "string",
                                        "maxLength": 150,
                                        "example": "Johan"
                                    },
                                    "email": {
                                        "description": "Correo del usuario (utilizado para el login)",
                                        "type": "string",
                                        "example": "correo@ejemplo.com"
                                    },
                                    "password": {
                                        "description": "Contraseña del usuario",
                                        "type": "string",
                                        "example": "contrasena123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transacción exitosa",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Acceso denegado"
                    },
                    "400": {
                        "description": "Petición no válida"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    },
                    "500": {
                        "description": "Error de servidor"
                    }
                }
            }
        },
        "/auth/usuarios/login": {
            "post": {
                "tags": [
                    "Usuarios - Autenticación"
                ],
                "summary": "Login",
                "description": "Servicio para la autenticación del usuario y la generación del token, que debe ser utilizado para consumir los Endpoints",
                "operationId": "usuarios/login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Correo registrado",
                                        "type": "string",
                                        "example": "correo@ejemplo.com"
                                    },
                                    "password": {
                                        "description": "Contraseña registrada",
                                        "type": "string",
                                        "example": "contrasena123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transacción exitosa",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Acceso denegado"
                    },
                    "400": {
                        "description": "Petición no válida"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    },
                    "500": {
                        "description": "Error de servidor"
                    }
                }
            }
        },
        "/auth/usuarios/info": {
            "get": {
                "tags": [
                    "Usuarios - Autenticación"
                ],
                "summary": "Información del usuario",
                "description": "Servicio para ver la información del usuario validado por el token que se está utilizando",
                "operationId": "usuarios/info",
                "requestBody": {},
                "responses": {
                    "200": {
                        "description": "Transacción exitosa",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Acceso denegado"
                    },
                    "400": {
                        "description": "Petición no válida"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    },
                    "500": {
                        "description": "Error de servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/auth/productos": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Listar productos",
                "description": "Servicio para visualizar todos los productos registrados en la base de datos",
                "operationId": "productos",
                "requestBody": {},
                "responses": {
                    "200": {
                        "description": "Transacción exitosa",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Acceso denegado"
                    },
                    "400": {
                        "description": "Petición no válida"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    },
                    "500": {
                        "description": "Error de servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}